Remove duplicate copies of remote code
