+++ esphome/components/remote_transmitter/__init__.py
@@ -131,8 +131,8 @@ FILTER_SOURCE_FILES = filter_source_files_from_platform(
             PlatformFramework.ESP32_ARDUINO,
             PlatformFramework.ESP32_IDF,
         },
-        "remote_transmitter_esp8266.cpp": {PlatformFramework.ESP8266_ARDUINO},
-        "remote_transmitter_libretiny.cpp": {
+        "remote_transmitter.cpp": {
+            PlatformFramework.ESP8266_ARDUINO,
             PlatformFramework.BK72XX_ARDUINO,
             PlatformFramework.RTL87XX_ARDUINO,
             PlatformFramework.LN882X_ARDUINO,
